feat: lotto 클래스 수정
lotto 클래스 내에 번호의 유효성 검사 로직을 추가하고 번호 출력 메서드, numbers에 접근하기 위한 getter를 추가합니다.
